Refactor ArchiveRoot
