+++ packages/client/modules/teamDashboard/containers/TeamArchive/TeamArchiveRoot.tsx
@@ -1,41 +1,16 @@
 import React from 'react'
-import graphql from 'babel-plugin-relay/macro'
 import {RouteComponentProps} from 'react-router'
-import {QueryRenderer} from 'react-relay'
-import TeamArchive from '../../components/TeamArchive/TeamArchive'
-import {LoaderSize} from '../../../../types/constEnums'
-import renderQuery from '../../../../utils/relay/renderQuery'
-import useAtmosphere from '../../../../hooks/useAtmosphere'
+import ArchiveTaskRoot, {ArchiveTaskRootProps} from '~/components/ArchiveTaskRoot'
 
-const query = graphql`
-  query TeamArchiveRootQuery($teamId: ID!, $first: Int!, $after: DateTime) {
-    viewer {
-      ...TeamArchive_viewer
-    }
-  }
-`
-
-interface Props extends RouteComponentProps<{teamId: string}> {
-  team: any
+interface Props extends RouteComponentProps<{teamId: string}>, ArchiveTaskRootProps {
 }
 
 const TeamArchiveRoot = ({match, team}: Props) => {
-  const atmosphere = useAtmosphere()
   const {
     params: {teamId}
   } = match
-  const {viewerId} = atmosphere
   return (
-    <QueryRenderer
-      environment={atmosphere}
-      query={query}
-      variables={{teamId, first: 40}}
-      fetchPolicy={'store-or-network' as any}
-      render={renderQuery(TeamArchive, {
-        props: {teamId, team, userId: viewerId, showHeader: true},
-        size: LoaderSize.PANEL
-      })}
-    />
+    <ArchiveTaskRoot teamId={teamId} team={team} showHeader={true} />
   )
 }
